refactor(web): start of OSKView abstraction
