+++ web/source/osk/oskManager.ts
@@ -1,15 +1,11 @@
-/// <reference path="../kmwexthtml.ts" />  // Includes KMW-added property declaration extensions for HTML elements.
+// Includes KMW-added property declaration extensions for HTML elements.
+/// <reference path="../kmwexthtml.ts" />
 // Includes the touch-mode language picker UI.
 /// <reference path="languageMenu.ts" />
-// Includes the banner
-/// <reference path="./bannerManager.ts" />
+/// <reference path="lengthStyle.ts" />
 // Defines desktop-centric OSK positioning + sizing behavior
 /// <reference path="layouts/targetedFloatLayout.ts" />
-// Generates the visual keyboard specific to each keyboard.  (class="kmw-osk-inner-frame")
-/// <reference path="visualKeyboard.ts" />
-// Models keyboards that present a help page, rather than a standard OSK.
-/// <reference path="helpPageView.ts" />
-/// <reference path="emptyView.ts" />
+/// <reference path="oskView.ts" />
 
 /***
    KeymanWeb 10.0
@@ -21,27 +17,7 @@ namespace com.keyman.osk {
     'nosize'?: boolean, 'nomove'?: boolean};
   type OSKPos = {'left'?: number, 'top'?: number};
 
-  export class OSKManager {
-    // Important OSK elements (and container classes)
-    _Box: HTMLDivElement;
-
-    headerView:   OSKViewComponent;
-    bannerView:   BannerManager; // Which implements OSKViewComponent
-    keyboardView: KeyboardView;  // Which implements OSKViewComponent
-    footerView:   OSKViewComponent;
-
-    public get vkbd(): VisualKeyboard {
-      if(this.keyboardView instanceof VisualKeyboard) {
-        return this.keyboardView;
-      } else {
-        return null;
-      }
-    }
-
-    public get banner(): BannerManager {  // Maintains old reference point used by embedding apps.
-      return this.bannerView;
-    }
-
+  export class OSKManager extends OSKView {
     desktopLayout: layouts.TargetedFloatLayout;
 
     // OSK state fields
@@ -57,38 +33,6 @@ namespace com.keyman.osk {
     dfltX: string;
     dfltY: string;
 
-    /**
-     * The configured width for this OSKManager.  May be `undefined` or `null`
-     * to allow automatic width scaling. 
-     */
-    private _width: ParsedLengthStyle;
-
-    /**
-     * The configured height for this OSKManager.  May be `undefined` or `null`
-     * to allow automatic height scaling. 
-     */
-    private _height: ParsedLengthStyle;
-
-    /**
-     * The computed width for this OSKManager.  May be null if auto sizing
-     * is allowed and the OSKManager is not currently in the DOM hierarchy.
-     */
-    private _computedWidth: number;
-
-    /**
-    * The computed height for this OSKManager.  May be null if auto sizing
-    * is allowed and the OSKManager is not currently in the DOM hierarchy.
-    */
-    private _computedHeight: number;
-
-    /**
-     * The base font size to use for hosted `Banner`s and `VisualKeyboard`
-     * instances.
-     */
-    private _baseFontSize: ParsedLengthStyle;
-
-    private needsLayout: boolean = true;
-
     // Key code definition aliases for legacy keyboards  (They expect window['keyman']['osk'].___)
     modifierCodes = text.Codes.modifierCodes;
     modifierBitmasks = text.Codes.modifierBitmasks;
@@ -96,6 +40,8 @@ namespace com.keyman.osk {
     keyCodes = text.Codes.keyCodes;
 
     public constructor() {
+      super();
+
       let keymanweb = com.keyman.singleton;
       let util = keymanweb.util;
 
@@ -232,8 +178,8 @@ namespace com.keyman.osk {
 
       // Initializes the size of a touch keyboard.
       if(this.vkbd && device.touchable) {
-        let targetOSKHeight = this.vkbd.computedAdjustedOskHeight(this.getKeyboardHeight());
-        this.setSize(this.getWidth(), targetOSKHeight + this.getBannerHeight());
+        let targetOSKHeight = this.vkbd.computedAdjustedOskHeight(this.getDefaultKeyboardHeight());
+        this.setSize(this.getWidth(), targetOSKHeight + this.banner.height);
       }
       // END:  construction of the actual internal layout for the overall OSK
 
@@ -262,90 +208,6 @@ namespace com.keyman.osk {
       if(this._Enabled) {
         this._Show();
       }
-    }
-
-    /**
-     * The configured width for this VisualKeyboard.  May be `undefined` or `null`
-     * to allow automatic width scaling. 
-     */
-     get width(): LengthStyle {
-      return this._width;
-    }
-
-    /**
-     * The configured height for this VisualKeyboard.  May be `undefined` or `null`
-     * to allow automatic height scaling. 
-     */
-    get height(): LengthStyle {
-      return this._height;
-    }
-
-    /**
-     * The computed width for this VisualKeyboard.  May be null if auto sizing
-     * is allowed and the VisualKeyboard is not currently in the DOM hierarchy.
-     */
-    get computedWidth(): number {
-      // Computed during layout operations; allows caching instead of continuous recomputation.
-      if(this.needsLayout) {
-        this.refreshLayout();
-      }
-      return this._computedWidth;
-    }
-
-    /**
-     * The computed height for this VisualKeyboard.  May be null if auto sizing
-     * is allowed and the VisualKeyboard is not currently in the DOM hierarchy.
-     */
-    get computedHeight(): number {
-      // Computed during layout operations; allows caching instead of continuous recomputation.
-      if(this.needsLayout) {
-        this.refreshLayout();
-      }
-      return this._computedHeight;
-    }
-
-    /**
-     * The top-level style string for the font size used by the predictive banner 
-     * and the primary keyboard visualization elements.
-     */
-    get baseFontSize(): string {
-      return this.parsedBaseFontSize.styleString;
-    }
-
-    private get parsedBaseFontSize(): ParsedLengthStyle {
-      if(!this._baseFontSize) {
-        let keymanweb = com.keyman.singleton;
-        let device = keymanweb.util.device;
-        this._baseFontSize = this.defaultFontSize(device, keymanweb.isEmbedded);
-      }
-
-      return this._baseFontSize;
-    }
-
-    public defaultFontSize(device: Device, isEmbedded: boolean): ParsedLengthStyle {
-      if(device.touchable) {
-        var fontScale: number = 1;
-        if(device.formFactor == 'phone') {
-          fontScale = 1.6 * (isEmbedded ? 0.65 : 0.6) * 1.2;  // Combines original scaling factor with one previously applied to the layer group.
-        } else {
-          // The following is a *temporary* fix for small format tablets, e.g. PendoPad
-          var pixelRatio = 1;
-          if(device.OS == 'Android' && 'devicePixelRatio' in window) {
-            pixelRatio = window.devicePixelRatio;
-          }
-
-          if(device.OS == 'Android' && device.formFactor == 'tablet' && this.getHeight() < 300 * pixelRatio) {
-            fontScale *= 1.2;
-          } else {
-            fontScale *= 2; //'2.5em';
-          }
-        }
-
-        // Finalize the font size parameter.
-        return ParsedLengthStyle.special(fontScale, 'em');
-      } else {
-        return this.computedHeight ? ParsedLengthStyle.inPixels(this.computedHeight / 8) : undefined;
-      } 
     }
 
     private layerChangeHandler: text.SystemStoreMutationHandler = function(this: OSKManager,
@@ -580,46 +442,11 @@ namespace com.keyman.osk {
       }
     }
 
-    /*private*/ setSize(width?: number, height?: number, pending?: boolean) {
-      let mutatedFlag = false;
-
-      if(width && height) {
-        mutatedFlag = !this._width || !this._height;
-
-        const parsedWidth = ParsedLengthStyle.inPixels(width);
-        const parsedHeight = ParsedLengthStyle.inPixels(height);
-
-        mutatedFlag = mutatedFlag || parsedWidth.styleString  != this._width.styleString;
-        mutatedFlag = mutatedFlag || parsedHeight.styleString != this._height.styleString;
-
-        this._width = parsedWidth;
-        this._height = parsedHeight;
-      }
-
-      this.needsLayout = this.needsLayout || mutatedFlag;
-
-      if(this.vkbd) {
-        let availableHeight = height - this.computeFrameHeight();
-        if(this.bannerView.height > 0) {
-          availableHeight -= this.bannerView.height + 5;
-        }
-        this.vkbd.setSize(width, availableHeight, pending);
-      }
-    }
-
-    /**
-     * Get the wanted height of the banner (does not include the keyboard)
-     *  @return   {number}    height in pixels
-     */
-    getBannerHeight(): number {
-      return (this.banner != null) ? this.banner.height : 0;
-    }
-
     /**
      * Get the wanted height of the OSK for touch devices (does not include banner height)
      *  @return   {number}    height in pixels
      **/
-    getKeyboardHeight(): number {
+    getDefaultKeyboardHeight(): number {
       let keymanweb = com.keyman.singleton;
       let device = keymanweb.util.device;
 
@@ -647,14 +474,6 @@ namespace com.keyman.osk {
       }
 
       return height;
-    }
-
-    /**
-     * Get the wanted height of the OSK for touch devices (banner height + rows of keys)
-     *  @return   {number}    height in pixels
-     **/
-    getHeight(): number {
-      return this.getBannerHeight() + this.getKeyboardHeight();
     }
 
     /**
@@ -859,10 +678,6 @@ namespace com.keyman.osk {
       }
     }
 
-    /* private */ computeFrameHeight(): number {
-      return (this.headerView?.layoutHeight.val || 0) + (this.footerView?.layoutHeight.val || 0);
-    }
-
     /**
      * Display KMW OSK at specified position (returns nothing)
      *
@@ -923,9 +738,8 @@ namespace com.keyman.osk {
       //TODO: may need to return here for touch devices??
       Ls.display='block'; //Ls.visibility='visible';
 
-      if(this.needsLayout) { // first thing after it's made visible.
-        this.refreshLayout();
-      }
+      // first thing after it's made visible.
+      this.refreshLayoutIfNeeded();
 
       if(this.vkbd) {
         this.vkbd.showLanguage();
@@ -991,45 +805,6 @@ namespace com.keyman.osk {
         Lpos['y']=this._Box.offsetTop;
         Lpos['userLocated']=this.userPositioned;
         this.doShow(Lpos);
-      }
-    }
-
-    public refreshLayout(): void {
-      // Step 1:  have the necessary conditions been met?
-      const fixedSize = this.width && this.height && this.width.absolute && this.height.absolute;
-      const computedStyle = getComputedStyle(this._Box);
-      const isInDOM = computedStyle.height != '' && computedStyle.height != 'auto';
-
-      // Step 2:  determine basic layout geometry
-      if(fixedSize) {
-        this._computedWidth  = this.width.val;
-        this._computedHeight = this.height.val;
-      } else if(isInDOM) {
-        this._computedWidth   = parseInt(computedStyle.width, 10);
-        this._computedHeight  = parseInt(computedStyle.height, 10);
-      } else {
-        // Cannot perform layout operations!
-        return;
-      }
-
-      // Must be set before any references to the .computedWidth and .computedHeight properties!
-      this.needsLayout = false;
-
-      // Step 3:  perform layout operations.
-      if(this.vkbd) {
-        let availableHeight = this.computedHeight - this.computeFrameHeight();
-
-        // +5:  from kmw-banner-bar's 'top' attribute when active
-        if(this.bannerView.height > 0) {
-          availableHeight -= this.bannerView.height + 5;
-        }
-        this.vkbd.setSize(this.computedWidth, availableHeight);
-        this.vkbd.refreshLayout();
-
-        if(this.vkbd.usesFixedHeightScaling) {
-          var b: HTMLElement = this._Box, bs=b.style;
-          bs.height=bs.maxHeight=this.computedHeight+'px';
-        }
       }
     }
